pushing non sorted elements to b

+++ t_stack/t_stack.h
@@ -6,7 +6,7 @@
 /*   By: ikhadem <ikhadem@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/28 14:59:04 by ikhadem           #+#    #+#             */
-/*   Updated: 2021/09/02 10:20:39 by ikhadem          ###   ########.fr       */
+/*   Updated: 2021/09/03 13:53:13 by ikhadem          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ t_stack		*t_stack_push(t_stack **, int);
 t_stack		*t_stack_pop(t_stack **);
 t_stack		*t_stack_copy(t_stack *self);
 size_t		t_stack_size(t_stack *);
-size_t		t_stack_index_of(t_stack *self, int key);
+int			t_stack_index_of(t_stack *self, int key);
 int			t_stack_top(t_stack **);
 int			t_stack_index(t_stack *, size_t);
 int			t_stack_max(t_stack *self);
